+++ tests/unit/modules/selenium/test_apiritif_builder.py
@@ -964,3 +964,21 @@ class TestApiritifScriptGeneration(ExecutorTestCase):
         self.obj.prepare()
         warnings = self.log_recorder.warn_buff.getvalue()
         self.assertIn("Passphrase was found, but certificate is missing!", warnings)
+
+    def test_stopping_reason_extracted(self):
+        concurrency = 4
+        iterations = 3
+
+        self.configure({"execution": {
+            "executor": "apiritif",
+            "concurrency": concurrency,
+            "iterations": iterations,
+            "scenario": {
+                "script": RESOURCES_DIR + "apiritif/test_stopping_reason.py"
+            }
+        }})
+        self.obj.prepare()
+        self.obj.startup()
+        self.obj.shutdown()
+        self.obj.post_process()
+        self.assertEqual(concurrency * iterations, len(self.obj.engine.extracted_stopping_reasons))
